Changes to the esp32 makefile to allow c++ modules to be built
I ran into problems where the c++ part of the build wouldn't work
when it had includes from py/obj.h or py/runtime.h.

The changes here resolved the issue.  I need to upstream them and see if
I can either get this project to reproduce the issue or alter the
example c++ module to throw the error.