ports/esp32: Setup usermod base files from rp2
We intend to use the usermod.cmake file provided by the rp2 port.

This commit holds the version of the file from
b64f29f233cfb8aebe0e0833a83032406fc727a1

made by:

Co-authored-by: Graham Sanderson <graham.sanderson@raspberrypi.org>
Signed-off-by: Phil Howard <phil@pimoroni.com>

It also includes the commit with the cmake files for the examples:

Add usermod.cmake files to the C and CPP examples.

Probably this commit can be removed once the first pull request is
merged to master.

At the moment it provides the common base file that the esp32 build will
use to implement user modules.

Signed-off-by: Michael O'Cleirigh <michael.ocleirigh@rivulet.ca>